+++ packages/spotlight/CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact spotlight module, newest
 ### Fixed
 
 - `spotlight/Spottable` to prevent unnecessary renders due to focus changes
+
 ### Deprecated
 
 ### Added